Test coverage improvements
